Access Google Maps. (0.1.0a)
Intializr was used to generate a template for this web application.
Icons are still default, and arbitrary plotting remains unimplemented.
Some files are still just templates and require more information to be
accurate. Google Analytics may be removed or properly supported in the
near future. Also, a mechanism is needed to retain an API key for
Google Maps across commits.