Implement coordinate retrieval. (0.2.0a)
A GeoJSON file can now be specified through the query string of the URL
and is retrieved using AJAX. Google Analytics code and the Google Maps
API key have been removed until further notice for the following
reasons: both will contain information specific to this application
(not the open source code) and neither will be useful until deployment.
Implementing a configuration file in JavaScript for this information
would require more AJAX which does not seem to be worthwhile. Also,
OpenStreetMap may replace Google Maps in the near future depending on
the applicability of available plotting libraries such as Leaflet.js.
Finally, only a single version of jQuery is needed, and it is going to
be stored locally. Code to include the same version from a remote
source has been removed.