GeoJSON file visualizations. (0.4.0a)
Vizit can now plot GeoJSON files. It uses the name property of GeoJSON
Feature objects to create a popup for each waypoint and allows points
to be customized by adding properties such as color and opacity (see
http://leafletjs.com/reference.html#path-options for details). It also
renders GeoJSON files with nested FeatureCollections, but they must
have a very specific format that specifies a single Feature as the
epicenter for an immediately proceeding FeatureCollection. One or many
of these sets must be contained in a wrapper FeatureCollection. An
example file has been included. Finally, a basic GeoJSON validator that
allows for nested FeatureCollections has been included; however, it
does not validate coordinates against corresponding type in Geometry
objects. This code may be moved to a separate file in the future. Also,
the map does not center itself appropriately yet.